removed createContractWeb3 route
